feat : add interface, modify login & logout logic and fix logout error
