+++ build.gradle
@@ -58,6 +58,7 @@ dependencies {
 	testImplementation 'org.springframework.security:spring-security-test'
 	implementation 'org.springframework.boot:spring-boot-starter-mustache'
 
+
 }
 
 tasks.named('test') {
