+++ src/main/java/com/hanshin/supernova/auth/dto/request/AuthLoginRequest.java
@@ -1,0 +1,9 @@
+package com.hanshin.supernova.auth.dto.request;
+
+import lombok.Data;
+
+@Data
+public class AuthLoginRequest {
+    private String email;
+    private String password;
+}
