+++ src/main/java/com/hanshin/supernova/auth/infrastructure/RefreshTokenRepository.java
@@ -1,0 +1,13 @@
+package com.hanshin.supernova.auth.infrastructure;
+
+import com.hanshin.supernova.auth.model.RefreshToken;
+import com.hanshin.supernova.user.domain.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
+    Optional<RefreshToken> findByToken(String token);
+    Optional<RefreshToken> findByUserId(Long userId);
+    void deleteByUserId(Long userId);
+}
