+++ src/main/java/com/hanshin/supernova/auth/model/AuthUser.java
@@ -1,0 +1,5 @@
+package com.hanshin.supernova.auth.model;
+
+public interface AuthUser {
+    Long getId();
+}
