--- src/main/java/com/hanshin/supernova/auth/v1/dto/request/AuthLoginRequest.java
@@ -1,10 +1,0 @@
-package com.hanshin.supernova.auth.v1.dto.request;
-
-import lombok.Data;
-
-@Data
-public class AuthLoginRequest {
-
-    private String email;
-    private String password;
-}
