--- src/main/java/com/hanshin/supernova/auth/v1/dto/response/AuthLoginResponse.java
@@ -1,12 +1,0 @@
-package com.hanshin.supernova.auth.v1.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class AuthLoginResponse {
-
-    private String nickname;
-    private String token;
-}
