--- src/main/java/com/hanshin/supernova/auth/v2/model/SecurityAuthUser.java
@@ -1,5 +1,0 @@
-package com.hanshin.supernova.auth.v2.model;
-
-public interface SecurityAuthUser {
-    Long getId();
-}
