+++ src/main/java/com/hanshin/supernova/news/presentation/NewsController.java
@@ -1,6 +1,6 @@
 package com.hanshin.supernova.news.presentation;
 
-import com.hanshin.supernova.auth.v2.model.SecurityAuthUser;
+import com.hanshin.supernova.auth.model.AuthUser;
 import com.hanshin.supernova.common.model.ResponseDto;
 import com.hanshin.supernova.news.application.NewsService;
 import com.hanshin.supernova.news.dto.request.NewsRequest;
@@ -38,7 +38,7 @@ public class NewsController {
      */
     @GetMapping(path = "/{newsId}")
     public ResponseEntity<?> getNews(
-            SecurityAuthUser user,
+            AuthUser user,
             @PathVariable(name = "newsId") Long newsId
     ) {
         var response = newsService.getNews(user, newsId);
@@ -50,7 +50,7 @@ public class NewsController {
      */
     @PutMapping(path = "/{newsId}")
     public ResponseEntity<?> updateNews(
-            SecurityAuthUser user,
+            AuthUser user,
             @PathVariable(name = "newsId") Long newsId,
             @RequestBody @Valid NewsRequest request
     ) {
@@ -63,7 +63,7 @@ public class NewsController {
      */
     @DeleteMapping(path = "/{newsId}")
     public ResponseEntity<?> deleteNews(
-            SecurityAuthUser user,
+            AuthUser user,
             @PathVariable(name = "newsId") Long newsId
     ) {
         var response = newsService.deleteNews(user, newsId);
@@ -75,7 +75,7 @@ public class NewsController {
      */
     @GetMapping(path = "/unviewed-list")
     public ResponseEntity<?> getUnViewedNews(
-            SecurityAuthUser user
+            AuthUser user
     ) {
         var responses = newsService.getUnViewedNews(user);
         return ResponseDto.ok(responses);
@@ -86,7 +86,7 @@ public class NewsController {
      */
     @GetMapping(path = "/viewed-list")
     public ResponseEntity<?> getViewedNews(
-            SecurityAuthUser user
+            AuthUser user
     ) {
         var responses = newsService.getViewedNews(user);
         return ResponseDto.ok(responses);
