+++ src/main/java/com/hanshin/supernova/user/application/UserService.java
@@ -1,0 +1,10 @@
+package com.hanshin.supernova.user.application;
+
+import com.hanshin.supernova.user.domain.User;
+import com.hanshin.supernova.user.dto.request.UserRegisterRequest;
+import com.hanshin.supernova.user.dto.response.UserRegisterResponse;
+
+public interface UserService {
+    UserRegisterResponse registerUser(UserRegisterRequest request);
+    User loadUserByEmail(String email);
+}
