+++ src/main/java/com/hanshin/supernova/user/domain/Authority.java
@@ -1,0 +1,6 @@
+package com.hanshin.supernova.user.domain;
+
+public enum Authority {
+
+    USER, ADMIN, SYSTEM
+}
