+++ src/main/java/com/hanshin/supernova/user/dto/response/UserRegisterResponse.java
@@ -1,0 +1,11 @@
+package com.hanshin.supernova.user.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class UserRegisterResponse {
+    private Long id;
+    private String nickname;
+}
