+++ src/main/java/com/hanshin/supernova/user/infrastructure/UserRepository.java
@@ -1,0 +1,19 @@
+package com.hanshin.supernova.user.infrastructure;
+
+import com.hanshin.supernova.user.domain.Authority;
+import com.hanshin.supernova.user.domain.User;
+import com.hanshin.supernova.user.domain.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+
+    boolean existsByEmail(String email);
+    boolean existsByNickname(String nickname);
+
+    Optional<User> findByEmail(String email);
+    Optional<User> findByAuthority(Authority authority);
+}
