+++ src/main/java/com/hanshin/supernova/user/presentation/UserController.java
@@ -1,0 +1,29 @@
+package com.hanshin.supernova.user.presentation;
+
+import com.hanshin.supernova.common.model.ResponseDto;
+import com.hanshin.supernova.user.application.UserService;
+import com.hanshin.supernova.user.dto.request.UserRegisterRequest;
+import com.hanshin.supernova.user.dto.response.UserRegisterResponse;
+import jakarta.validation.Valid;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping(path = "/api/users", produces = MediaType.APPLICATION_JSON_VALUE)
+@RequiredArgsConstructor
+public class UserController {
+
+    private final UserService userService;
+
+    @PostMapping("/register")
+    public ResponseEntity<UserRegisterResponse> register(@RequestBody @Valid UserRegisterRequest request){
+        UserRegisterResponse response = userService.registerUser(request);
+        return ResponseEntity.ok(response);
+    }
+
+}
