--- src/main/java/com/hanshin/supernova/user/v1/dto/response/UserRegisterResponse.java
@@ -1,12 +1,0 @@
-package com.hanshin.supernova.user.v1.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class UserRegisterResponse {
-
-    private Long id;
-    private String nickname;
-}
