--- src/main/java/com/hanshin/supernova/user/v2/domain/Authority.java
@@ -1,6 +1,0 @@
-package com.hanshin.supernova.user.v2.domain;
-
-public enum Authority {
-
-    USER, ADMIN, SYSTEM
-}
